data-population/pyproject.toml
[tool.ruff.lint]
select = ["ALL"]

[tool.ruff]
line-length = 80

[tool.black]
line-length = 80

# Ignore some errors in tests as they are not relevant
[tool.ruff.lint.extend-per-file-ignores]
"**/tests/**/*.py" = [
    # at least this three should be fine in tests:
    "S101", # asserts allowed in tests...
    "S105", # local dev password fields are verified
    "ARG", # Unused function args -> fixtures nevertheless are functionally relevant...
    "FBT", # Don't care about booleans as positional arguments in tests, e.g. via @pytest.mark.parametrize()
    # The below are debateable
    "PLR2004", # Magic value used in comparison, ...
    "S311", # Standard pseudo-random generators are not suitable for cryptographic purposes
]

[project]
authors = [
    {name = "Your Name", email = "you@example.com"},
]
license = {text = "Apache License 2.0"}
requires-python = "<4.0,>=3.11"
dependencies = [
    "oracledb<3.0.0,>=2.4.1",
    "numpy<3.0.0,>=2.1.1",
    "SQLAlchemy<3.0.0,>=2.0.35",
    "pyyaml>=6.0.2,<7.0.0",
    "pandas<3.0.0,>=2.2.3",
    "pyarrow>=18.0.0,<19.0.0",
    "pypika<1.0.0,>=0.48.9",
    "boto3<2.0.0,==1.35.95",
    "kubernetes<32.0.0,>=31.0.0",
    "requests<3.0.0,>=2.32.3",
    "openshift-client<3.0.0,>=2.0.4",
    "kr8s<1.0.0,>=0.18.0",
    "psycopg2-binary<3.0.0,>=2.9.10",
    "fastparquet>=2024.11.0",
    "click>=8.1.8",
    "geopandas>=1.0.1",
    "fiona>=1.10.1",
]
name = "data-population"
version = "0.1.0"
description = "code used to extract and load data to local dev env"
readme = "README.md"
package-mode = false

[dependency-groups]
dev = [
    "black<25.0.0,>=24.8.0",
    "ruff<1.0.0,>=0.8.0",
    "mypy-boto3<2.0.0,>=1.35.38",
    "pytest<9.0.0,>=8.3.3",
]
